test: fix the width tests' table lookup and the CLI's type expectations
Two follow-ups from the engine run:

The catalog lookup guessed the table name. twAutoPK pluralises to
tw_auto_p_ks, not tw_auto_pks, so the column was never found on the two
engines that take that path — the metadata knows the name, so it is asked.

The CLI's from-models golden pinned SERIAL and REAL, which are exactly the two
types QK-21 was about: four-byte key, single-precision float.

Refs A4/S4.

+++ cmd/quark/commands/from_models_test.go
@@ -89,11 +89,13 @@ func TestMigrateCreateFromModelsPostgres(t *testing.T) {
 
 	for _, want := range []string{
 		`CREATE TABLE IF NOT EXISTS users`,
-		`SERIAL PRIMARY KEY`,
+		// 64-bit key and 64-bit float (QK-21): SERIAL is four bytes, and
+		// PostgreSQL's REAL is single precision.
+		`BIGSERIAL PRIMARY KEY`,
 		`VARCHAR(190) UNIQUE`,
 		`NOT NULL`,
 		`CREATE TABLE IF NOT EXISTS articles`,
-		`score REAL`,
+		`score DOUBLE PRECISION`,
 		`BOOLEAN`,
 		`REFERENCES users(id)`,
 		`CREATE INDEX IF NOT EXISTS idx_articles_author_id`,
